refactor: move Firebase setups to index.helper.ts
